+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog for [squel](https://github.com/hiddentao/squel)
 
+## 28 Feb 2017 (5.8.0)
+* #301 - Add `rstr()` to enable "raw" nesting of query builders
+* Renamed `_isSquelBuilder()` call to `isSquelBuilder()`
+
 ## 6 Feb 2017 (5.7.0)
 * #288 - more flexible `RETURNING` clauses